bot/pyproject.toml
[project]
name = "bot"
version = "1.0.0"
requires-python = ">=3.11"
description = ""
dependencies = [
    "discord-py>=2.4.0",
    "motor>=3.6.0",
    "pydantic>=2.9.2",
    "pydantic-settings>=2.5.2",
    "pyyaml>=6.0.2",
]
